Fold the sgx-hasher crate into the sgx-types crate
This should provide a more natural API.

We hide the crypto code behind the "openssl" feature to avoid forcing a
dependency on all consumers.
