+++ Cargo.toml
@@ -3,5 +3,4 @@ members = [
     "sgx-types",
     "sgx-traits",
     "sgx-sys",
-    "sgx-hasher",
 ]
