--- sgx-hasher/Cargo.toml
@@ -1,11 +1,0 @@
-[package]
-name = "sgx-hasher"
-version = "0.1.0"
-authors = ["Lily Sturmann <lsturman@redhat.com>", "Nathaniel McCallum <npmccallum@redhat.com>" ]
-edition = "2018"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-sgx-types = { path = "../sgx-types" }
-openssl = "0.10.26"
