+++ sgx-types/Cargo.toml
@@ -6,5 +6,9 @@ edition = "2018"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
+[features]
+default = ["openssl"]
+
 [dependencies]
+openssl = { version = "0.10.26", optional = true }
 bitflags = "1.2.1"
