+++ sgx-types/tests/encl.bin
